code-plagiarism-detection/requirements.txt
requests
grequests

PyYAML
autopep8

xlwt
